requirements.txt
Flask=>2.0.2
Flask-Session=>0.4.0
Bootstrap-Flask=>1.8.0
PyYAML=>6.0
OpenVAS-Reporting=>1.5.2
defusedxml>=0.7.1
